backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
httpx==0.27.2
